requirements.txt
pydantic==2.11.10
aiofiles==24.1.0
tornado==6.5.2
httpx==0.28.1
lxml==6.0.1
PyYAML==6.0.2
psutil==7.0.0
pillow==11.3.0
pystray==0.19.5
loguru==0.7.3
APScheduler==3.11.0
pyinstaller==6.15.0
SQLAlchemy==2.0.43
# alembic==1.16.5
redis==6.4.0
# psycopg2-binary==2.9.10
# kafka-python==2.2.15